Main pr fixes
* refactoring part 1

* expires at

* fixes

* rm pointers

* split up services

* int64 for userID

* split providers

* retry http

* float to int

* float to int

* balance update into db transactions

* process orders

+++ api/rest/handler/balance.go
@@ -13,7 +13,7 @@ import (
 func (h Handler) ShowBalance(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 
-	userID, ok := ctx.Value(contexts.ContextUserKey).(int)
+	userID, ok := ctx.Value(contexts.ContextUserKey).(uint64)
 	if !ok {
 		http.Error(w, "failing with user context", http.StatusInternalServerError)
 		return
@@ -46,7 +46,7 @@ func (h Handler) ShowBalance(w http.ResponseWriter, r *http.Request) {
 func (h Handler) Withdraw(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 
-	userID, ok := ctx.Value(contexts.ContextUserKey).(int)
+	userID, ok := ctx.Value(contexts.ContextUserKey).(uint64)
 	if !ok {
 		http.Error(w, "failing with user context", http.StatusInternalServerError)
 		return
@@ -85,7 +85,7 @@ func (h Handler) Withdraw(w http.ResponseWriter, r *http.Request) {
 func (h Handler) Withdrawals(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 
-	userID, ok := ctx.Value(contexts.ContextUserKey).(int)
+	userID, ok := ctx.Value(contexts.ContextUserKey).(uint64)
 	if !ok {
 		http.Error(w, "failing with user context", http.StatusInternalServerError)
 		return
