+++ model/user.go
@@ -2,10 +2,10 @@ package model
 
 // User model.
 type User struct {
-	ID            int
+	ID            uint64
 	Login         string
 	Password      string
 	CryptPassword string
-	Balance       float64
-	Outcome       float64
+	Balance       int
+	Outcome       int
 }
